requirements.txt
Django>=3.2.3,<3.3

django-mptt==0.13.4

psycopg2>=2.8.6,<2.9
psycopg2-binary==2.9.1

celery==5.1.2
django-celery-results==2.2.0
redis==3.5.3

uWSGI>=2.0.19.1,<2.1

django-health-check==3.16.5
Pillow==9.1.0
django-hitcount==1.3.5
sorl-thumbnail==12.8.0